+++ lib/angelo.rb
@@ -24,6 +24,7 @@ module Angelo
   DEFAULT_CONTENT_DISPOSITION = 'attachment'
   ETAG_HEADER_KEY = 'ETag'
   IF_NONE_MATCH_HEADER_KEY = 'If-None-Match'
+  LOCATION_HEADER_KEY = 'Location'
 
   HTML_TYPE = 'text/html'
   JSON_TYPE = 'application/json'